adds NPC folder, fixed namespace issue
